requirements.txt
flask>=2.2.5
mysql-connector-python==8.0.26
flask-cors>=4.0.0
python-dotenv>=1.0.0
